please tell me this works
having troubles with git but: svm and svr are now free of errors. accuracy isn't really groundbreaking and is kind of all over the place, but it's worth toying with (especially region 1 in svr). dbscan code created but it's not going so well (zero clusters and like 2000 points of noise??). cnn created but not started yet. to do: possibly look into dbscan and figure out how to improve code. start on cnn. good news is that conda is now working on this computer!