updated svr, svm, dbscan (see descrip)
false hope yesterday: conda isn't working again :( so I can't work on cnn here. But tried reducing the number of 0s in svm to reduce skew/bias and accuracy went down as expected, but at least it'll probably be more realistic. dbscan is kind of a struggle because of euclidean distances and stuff. only good results when passing in two variables, and then the clusters are kind of chaotic, but it's something to go on. to do: try different combinations of variables in svr, svm, dbscan and see if anything interesting comes up, and also work on cnn at home