ugh (see descrip)
wildfire_cnn uploaded, but it's not going great. constant errors with the size of input and output when trying to train the model. code works up until there, but training won't work because the dimensions of the data is wrong? (asking for 3 dimensions but the way I did it apparently doesn't work?) to do: try to fix the dimensions and continue testing other programs, maybe take a look at the other models i tried out and improve the code in them