continued work on roc/pr
kept trying new pr/roc functions; the results of the functions keep changing whenever I change something slightly - pr is still a straight line but at some point it became a legit pr graph, but i'm not sure what i did differently. roc works when predict is not 0/1 (but it throws warnings about creating thresholds), but doesn't work when predict is 0/1. to do: figure out how to get the legit pr graph back and find a way to have roc with 0/1 data.